Don't die on syntax errors
QA now treats syntax errors like test errors. This
makes it easier to use on a code base that's
actively being edited. This behavior will be most
useful when combined with (still experimental)
`qa auto` functionality.

Also fix bad timestamps in TAPJ stream. On some
versions of Ruby, we use a monotonic clock that
needs processing before it can be converted to
a date.